% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_regulation_matrices.r
\name{calculate.regulation.matrices}
\alias{calculate.regulation.matrices}
\title{Calculate regulation matrices}
\usage{
calculate.regulation.matrices(rg, cache.file.path, cache.ignore, alpha = 1)
}
\arguments{
\item{rg}{An \link{igraph} graph object with the regulation graph, usually
obtained with \code{\link[=build.regulation.graph]{build.regulation.graph()}}.}

\item{cache.file.path}{Character string with the directory and name of an R
data file to save and reuse the calculated matrices.}

\item{cache.ignore}{Boolean indicating whether to force execution and
ignore a possible previous cache.}
}
\value{
A list with two elements:
\itemize{
\item \code{step}: Matrix with one-step regulation.
\item \code{flow}: Matrix with all-step regulation.
}
}
\description{
It calculates one-step and all-step regulation matrices from a regulation
graph.
}
