% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_interactions.r
\name{get.interactions}
\alias{get.interactions}
\title{Get interactions}
\usage{
get.interactions(
  database.name,
  database.file.path,
  table.file.path,
  cache.file.path,
  cache.ignore
)
}
\arguments{
\item{database.name}{Character string with the name of the database to use.
Currently supported databases are given by \code{\link[=get.supported.databases]{get.supported.databases()}}.}

\item{database.file.path}{Character string with the directory and name of an
R object with interactions imported from the database. This file will
usually have been generated with the corresponding function
\verb{import.database.*()}.}

\item{table.file.path}{Character string with the directory and name of a CSV
table to save processed interactions, with both standard and original
format.}

\item{cache.file.path}{Character string with the directory and name of an R
data file to save and reuse processed interactions.}

\item{cache.ignore}{Boolean indicating whether to force execution and
ignore a possible previous cache.}
}
\value{
A data frame with the processed interactions in standard format.
}
\description{
It gets interactions from a database, by reading them from an R file
imported from the database. Interactions are selected and formatted with
the standard interaction names as given by
\code{\link[=get.standard.names.in.interactions]{get.standard.names.in.interactions()}}. A CSV table is generated as
reference for the selected interactions. A cache is also generated, for
optional reuse in later executions.
}
