% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_standard_names_in_interactions.r
\name{get.standard.names.in.interactions}
\alias{get.standard.names.in.interactions}
\title{Get standard names in interactions}
\usage{
get.standard.names.in.interactions()
}
\value{
Vector of character strings.
}
\description{
It returns a vector of character strings with the standard names in
interactions.
}
