% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_supported_databases.r
\name{get.supported.databases}
\alias{get.supported.databases}
\title{Get supported databases}
\usage{
get.supported.databases()
}
\value{
Vector of character strings.
}
\description{
It returns a vector of character strings with the names of the supported
databases, in alphabetical order.
}
