% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_database_omnipath.r
\name{import.database.omnipath}
\alias{import.database.omnipath}
\title{Import database from OmniPath}
\usage{
import.database.omnipath(database.file.path, only.protein.coding)
}
\arguments{
\item{database.file.path}{Character string with the directory and name of
the file to save the selected interactions.}

\item{only.protein.coding}{Boolean indicating whether to only select
interactions involving protein-coding genes.}
}
\value{
\code{NULL}, invisibly.
}
\description{
It downloads the interaction database from the OmniPath server, using the
library \link{OmnipathR}. It saves to a data file the subset of interactions
involving genes with HUGO names, including complexes. This data file will
usually be read when calling \code{\link[=get.interactions]{get.interactions()}}.
}
