% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_joint_regulation_flow.r
\name{plot_joint.regulation.flow}
\alias{plot_joint.regulation.flow}
\title{Plot joint values of regulation flow for two genes}
\usage{
plot_joint.regulation.flow(
  regulation.flow,
  gene.from,
  gene.to,
  is.from,
  plot.dir
)
}
\arguments{
\item{regulation.flow}{A matrix with regulation flow values, obtained with
\code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{gene.from}{Character vector with gene names as source of regulation.}

\item{gene.to}{Character vector with gene names as target of regulation.}

\item{is.from}{Boolean indicating:
\itemize{
\item \code{TRUE}: Downstream regulation from the two genes in \code{gene.from}.
\item \code{FALSE}: Upstream regulation to the two genes in \code{gene.to}.
}}

\item{plot.dir}{Character string with the directory to save the plot.}
}
\value{
\code{NULL}.
}
\description{
It obtains scatter plots of regulation flow values, separated by synergistic
and antagonistic effects, positive or negative. It overlays gene labels to
highlight particular set of genes.
}
\details{
The call is ignored if \code{gene.from} or \code{gene.to} (depending on \code{is.from} is
\code{TRUE} or \code{FALSE}) is \code{NULL} or with length different from 2.
}
