% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pathway.r
\name{plot_pathway}
\alias{plot_pathway}
\title{Plot pathway representations}
\usage{
plot_pathway(
  rg,
  regulation.step,
  regulation.flow,
  gene.from,
  gene.to,
  plot.dir,
  plot.file.name.base = NULL,
  plot.title.base = NULL,
  pathway.regulation.approx = 0.05,
  pathway.n.max = 10,
  pathway.gene.max = 20
)
}
\arguments{
\item{rg}{An \link{igraph} graph object with the interaction graph, usually
obtained \code{\link[=build.regulation.graph]{build.regulation.graph()}}.}

\item{regulation.step}{A matrix with one-step regulation flow values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{regulation.flow}{A matrix with all-step regulation flow values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{gene.from}{Character vector with gene names as source of regulation.}

\item{gene.to}{Character vector with gene names as target of regulation.
Elements also in \code{gene.from} are ignored.}

\item{plot.dir}{Character string with the directory to save the plot.}

\item{plot.file.name.base}{Character string with a prefix for the plot file
names. If \code{NULL}, default values are generated depending on the number
of genes in \code{gene.from} and \code{gene.to}.}

\item{plot.title.base}{Character string with a prefix for the plot titles.
If \code{NULL}, default values are generated depending on the number of genes
in \code{gene.from} and \code{gene.to}.}

\item{pathway.regulation.approx}{Float with the desired approximation of
pathway coverage.}

\item{pathway.n.max}{Integer with the maximum number of pathways to plot.}

\item{pathway.gene.max}{Integer with the maximum number of genes to be
plotted in any pathway.}
}
\value{
\code{NULL}, invisibly.
}
\description{
It plots partial pathway representations between genes or set of genes, in
increasing amounts of genes included and flow explained.
}
\details{
The call is ignored if \code{gene.from} or \code{gene.to}, after removing overlapping
values from \code{gene.to}, is \code{NULL} or has no elements.
}
