% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regflow_package.r
\docType{package}
\name{regflow-package}
\alias{regflow}
\alias{regflow-package}
\title{regflow: A package for carrying out Regulatory Flow Analysis}
\description{
The \code{regflow} package provides utilities to build regulation graphs,
obtain regulation matrices, perform calculations about regulation flow
and pathways, and generate figures with flow distributions and
representations of pathways.
}
