% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_pathway.r
\name{write.pathway}
\alias{write.pathway}
\title{Write table with regulation of intermediate genes in a pathway}
\usage{
write.pathway(
  regulation.flow,
  gene.from,
  gene.to,
  write.dir,
  write.file.name = NULL
)
}
\arguments{
\item{regulation.flow}{A matrix with all-step regulation flow values,
usually obtained with \code{\link[=calculate.regulation.matrices]{calculate.regulation.matrices()}}.}

\item{gene.from}{Character vector with gene names as source of regulation.}

\item{gene.to}{Character vector with gene names as target of regulation.
Elements also in \code{gene.from} are ignored.}

\item{write.dir}{Character string with the directory to write the table.}

\item{write.file.name}{Character string with the table file name. If \code{NULL},
default values are generated depending on the number of genes in
\code{gene.from} and \code{gene.to}.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Writes a CSV table with the amount of regulation that flows through
intermediate genes or proteins in a pathway. Values are given as absolute
values and relative values with respect to the overall regulation from
\code{gene.from} to \code{gene.to}.
}
\details{
The call is ignored if \code{gene.from} or \code{gene.to}, after removing overlapping
values from \code{gene.to}, is \code{NULL} or has no elements.
}
